A tidy implementation of the summarised_result object for incidence results.

A tidy implementation of the summarised_result object for incidence results.

Usage

asIncidenceResult(result, metadata = FALSE)

Arguments

result

A summarised_result object created by the IncidencePrevalence package.

metadata

If TRUE additional metadata columns will be included in the result.

Value

A tibble with a tidy version of the summarised_result object.

Examples

# \donttest{
cdm <- mockIncidencePrevalence()
inc <- estimateIncidence(cdm, "target", "outcome")
#> ℹ Getting incidence for analysis 1 of 1
#> ✔ Overall time taken: 0 mins and 1 secs
tidy_inc <- asIncidenceResult(inc)
# }