Run benchmark of incidence and prevalence analyses

Run benchmark of incidence and prevalence analyses

Usage

benchmarkIncidencePrevalence(
  cdm,
  cohortDateRange = as.Date(c(NA, NA)),
  returnParticipants = FALSE,
  nOutcomes = 1,
  prevOutcomes = 0.25,
  analysisType = "all",
  outputFolder = NULL,
  fileName = NULL
)

Arguments

cdm

A CDM reference object

cohortDateRange

Two dates. The first indicating the earliest cohort start date and the second indicating the latest possible cohort end date. If NULL or the first date is set as missing, the earliest observation_start_date in the observation_period table will be used for the former. If NULL or the second date is set as missing, the latest observation_end_date in the observation_period table will be used for the latter.

returnParticipants

Whether to return participants (requires temporary to be FALSE)

nOutcomes

An integer specifying the number of outcomes to create in the denominator cohort

prevOutcomes

An array of integers for the prevalence of the outcomes in the population (in %). If the user wants all the outcomes with the same prevalence, they can also provide a single integer

analysisType

A string of the following: "all", "only incidence", "only prevalence"

outputFolder

Folder to save results as CSV

fileName

Name given to CSV with results

Value

a tibble with time taken for different analyses

Examples

# \donttest{
cdm <- mockIncidencePrevalenceRef(
  sampleSize = 100,
  earliestObservationStartDate = as.Date("2010-01-01"),
  latestObservationStartDate = as.Date("2010-01-01"),
  minDaysToObservationEnd = 364,
  maxDaysToObservationEnd = 364,
  outPre = 0.1
)

timings <- benchmarkIncidencePrevalence(cdm,
                                        nOutcomes = 2,
                                        prevOutcomes = c(0.1),
                                        analysisType = "only incidence"
)
#> Loading required namespace: testthat
#> ℹ Creating denominator cohorts
#> ✔ Cohorts created in 0 min and 6 sec
#> Getting incidence for analysis 1 of 64
#> Getting incidence for analysis 2 of 64
#> Getting incidence for analysis 3 of 64
#> Getting incidence for analysis 4 of 64
#> Getting incidence for analysis 5 of 64
#> Getting incidence for analysis 6 of 64
#> Getting incidence for analysis 7 of 64
#> Getting incidence for analysis 8 of 64
#> Getting incidence for analysis 9 of 64
#> Getting incidence for analysis 10 of 64
#> Getting incidence for analysis 11 of 64
#> Getting incidence for analysis 12 of 64
#> Getting incidence for analysis 13 of 64
#> Getting incidence for analysis 14 of 64
#> Getting incidence for analysis 15 of 64
#> Getting incidence for analysis 16 of 64
#> Getting incidence for analysis 17 of 64
#> Getting incidence for analysis 18 of 64
#> Getting incidence for analysis 19 of 64
#> Getting incidence for analysis 20 of 64
#> Getting incidence for analysis 21 of 64
#> Getting incidence for analysis 22 of 64
#> Getting incidence for analysis 23 of 64
#> Getting incidence for analysis 24 of 64
#> Getting incidence for analysis 25 of 64
#> Getting incidence for analysis 26 of 64
#> Getting incidence for analysis 27 of 64
#> Getting incidence for analysis 28 of 64
#> Getting incidence for analysis 29 of 64
#> Getting incidence for analysis 30 of 64
#> Getting incidence for analysis 31 of 64
#> Getting incidence for analysis 32 of 64
#> Getting incidence for analysis 33 of 64
#> Getting incidence for analysis 34 of 64
#> Getting incidence for analysis 35 of 64
#> Getting incidence for analysis 36 of 64
#> Getting incidence for analysis 37 of 64
#> Getting incidence for analysis 38 of 64
#> Getting incidence for analysis 39 of 64
#> Getting incidence for analysis 40 of 64
#> Getting incidence for analysis 41 of 64
#> Getting incidence for analysis 42 of 64
#> Getting incidence for analysis 43 of 64
#> Getting incidence for analysis 44 of 64
#> Getting incidence for analysis 45 of 64
#> Getting incidence for analysis 46 of 64
#> Getting incidence for analysis 47 of 64
#> Getting incidence for analysis 48 of 64
#> Getting incidence for analysis 49 of 64
#> Getting incidence for analysis 50 of 64
#> Getting incidence for analysis 51 of 64
#> Getting incidence for analysis 52 of 64
#> Getting incidence for analysis 53 of 64
#> Getting incidence for analysis 54 of 64
#> Getting incidence for analysis 55 of 64
#> Getting incidence for analysis 56 of 64
#> Getting incidence for analysis 57 of 64
#> Getting incidence for analysis 58 of 64
#> Getting incidence for analysis 59 of 64
#> Getting incidence for analysis 60 of 64
#> Getting incidence for analysis 61 of 64
#> Getting incidence for analysis 62 of 64
#> Getting incidence for analysis 63 of 64
#> Getting incidence for analysis 64 of 64
#> Overall time taken: 0 mins and 26 secs
#> Getting incidence for analysis 1 of 64
#> Getting incidence for analysis 2 of 64
#> Getting incidence for analysis 3 of 64
#> Getting incidence for analysis 4 of 64
#> Getting incidence for analysis 5 of 64
#> Getting incidence for analysis 6 of 64
#> Getting incidence for analysis 7 of 64
#> Getting incidence for analysis 8 of 64
#> Getting incidence for analysis 9 of 64
#> Getting incidence for analysis 10 of 64
#> Getting incidence for analysis 11 of 64
#> Getting incidence for analysis 12 of 64
#> Getting incidence for analysis 13 of 64
#> Getting incidence for analysis 14 of 64
#> Getting incidence for analysis 15 of 64
#> Getting incidence for analysis 16 of 64
#> Getting incidence for analysis 17 of 64
#> Getting incidence for analysis 18 of 64
#> Getting incidence for analysis 19 of 64
#> Getting incidence for analysis 20 of 64
#> Getting incidence for analysis 21 of 64
#> Getting incidence for analysis 22 of 64
#> Getting incidence for analysis 23 of 64
#> Getting incidence for analysis 24 of 64
#> Getting incidence for analysis 25 of 64
#> Getting incidence for analysis 26 of 64
#> Getting incidence for analysis 27 of 64
#> Getting incidence for analysis 28 of 64
#> Getting incidence for analysis 29 of 64
#> Getting incidence for analysis 30 of 64
#> Getting incidence for analysis 31 of 64
#> Getting incidence for analysis 32 of 64
#> Getting incidence for analysis 33 of 64
#> Getting incidence for analysis 34 of 64
#> Getting incidence for analysis 35 of 64
#> Getting incidence for analysis 36 of 64
#> Getting incidence for analysis 37 of 64
#> Getting incidence for analysis 38 of 64
#> Getting incidence for analysis 39 of 64
#> Getting incidence for analysis 40 of 64
#> Getting incidence for analysis 41 of 64
#> Getting incidence for analysis 42 of 64
#> Getting incidence for analysis 43 of 64
#> Getting incidence for analysis 44 of 64
#> Getting incidence for analysis 45 of 64
#> Getting incidence for analysis 46 of 64
#> Getting incidence for analysis 47 of 64
#> Getting incidence for analysis 48 of 64
#> Getting incidence for analysis 49 of 64
#> Getting incidence for analysis 50 of 64
#> Getting incidence for analysis 51 of 64
#> Getting incidence for analysis 52 of 64
#> Getting incidence for analysis 53 of 64
#> Getting incidence for analysis 54 of 64
#> Getting incidence for analysis 55 of 64
#> Getting incidence for analysis 56 of 64
#> Getting incidence for analysis 57 of 64
#> Getting incidence for analysis 58 of 64
#> Getting incidence for analysis 59 of 64
#> Getting incidence for analysis 60 of 64
#> Getting incidence for analysis 61 of 64
#> Getting incidence for analysis 62 of 64
#> Getting incidence for analysis 63 of 64
#> Getting incidence for analysis 64 of 64
#> Overall time taken: 0 mins and 29 secs
# }