Identify a set of denominator populations

generateDenominatorCohortSet() creates a set of cohorts that can be used for the denominator population in analyses of incidence, using estimateIncidence(), or prevalence, using estimatePointPrevalence() or estimatePeriodPrevalence().

Usage

generateDenominatorCohortSet(
  cdm,
  startDate = NULL,
  endDate = NULL,
  ageGroup = list(c(0, 150)),
  sex = "Both",
  daysPriorHistory = 0,
  strataTable = NULL,
  strataCohortId = NULL,
  strataCohortName = NULL,
  sample = NULL,
  tablePrefix = NULL,
  verbose = FALSE
)

Arguments

cdm

A CDM reference object

startDate

A date indicating the start of the study period. If NULL, the earliest observation_start_date in the observation_period table will be used.

endDate

A date indicating the end of the study period. If NULL, the latest observation end date in the observation period table will be used.

ageGroup

A list of age groups for which cohorts will be generated. A value of list(c(0,17), c(18,30)) would, for example, lead to the creation of cohorts for those aged from 0 to 17 (up to the day before their 18th birthday), and from 18 (starting the day of their 18th birthday) to 30 (up to the day before their 31st birthday).

sex

Sex of the cohorts. This can be one or more of: "Male", "Female", or "Both".

daysPriorHistory

The number of days of prior history observed in the database required for an individual to start contributing time in a cohort.

strataTable

A cohort table in the cdm reference to use to limit cohort entry and exit (with individuals only contributing to a cohort when they are contributing to the cohort in the strata table).

strataCohortId

The cohort definition id for the cohort of interest in the strata table. Only one stratifying cohort is supported.

strataCohortName

Corresponding name for the strata cohort.

sample

An integer for which to take a random sample, using dplyr::slice_sample(), of the people in the person table eligible to be included in the cohort set.

tablePrefix

The stem for the permanent tables that will be created when creating the denominator cohorts. Permanent tables will be created using this stem, and any existing tables that start with this will be at risk of being dropped or overwritten. If NULL, temporary tables will be used throughout.

verbose

Either TRUE or FALSE. If TRUE, progress will be reported.

Value

A cohort reference

Examples

if (FALSE) {
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdm_from_con(
  con = con,
  cdm_schema = "main"
)

dpop <- generateDenominatorCohortSet(
  cdm = cdm,
  startDate = as.Date("2008-01-01"),
  endDate = as.Date("2018-01-01")
)
}