Table of prevalence attrition results

Table of prevalence attrition results

Usage

tablePrevalenceAttrition(
  result,
  type = "gt",
  header = "variable_name",
  groupColumn = c("cdm_name", "variable_level"),
  settingsColumns = colnames(settings(result)),
  hide = "estimate_name"
)

Arguments

result

A summarised_result object. Output of summariseCohortAttrition().

type

Type of table. Check supported types with visOmopResults::tableType().

header

Columns to use as header. See options with colnames(visOmopResults::splitAll(result)). Variables in settingsColumns are also allowed

groupColumn

Variables to use as group labels. Allowed columns are the same as in header

settingsColumns

Variables from the settings atribute to dispaly in the table

hide

Table columns to exclude, options are the ones described in header

Value

A visual table.