Run benchmark of incidence and prevalence analyses

Run benchmark of incidence and prevalence analyses

Usage

benchmarkIncidencePrevalence(
  cdm,
  cohortDateRange = NULL,
  temporary = TRUE,
  returnParticipants = FALSE,
  nOutcomes = 1,
  prevOutcomes = 0.25,
  analysisType = "all",
  outputFolder = NULL,
  fileName = NULL
)

Arguments

cdm

A CDM reference object

cohortDateRange

Two dates. The first indicating the earliest cohort start date and the second indicating the latest possible cohort end date. If NULL or the first date is set as missing, the earliest observation_start_date in the observation_period table will be used for the former. If NULL or the second date is set as missing, the latest observation_end_date in the observation_period table will be used for the latter.

temporary

If TRUE, temporary tables will be used throughout. If FALSE, permanent tables will be created in the write_schema of the cdm using the write_prefix (if specified). Note existing permanent tables in the write schema starting with the write_prefix will be at risk of being dropped or overwritten.

returnParticipants

Whether to return participants (requires temporary to be FALSE)

nOutcomes

An integer specifying the number of outcomes to create in the denominator cohort

prevOutcomes

An array of integers for the prevalence of the outcomes in the population (in %). If the user wants all the outcomes with the same prevalence, they can also provide a single integer

analysisType

A string of the following: "all", "only incidence", "only prevalence"

outputFolder

Folder to save results as CSV

fileName

Name given to CSV with results

Value

a tibble with time taken for different analyses

Examples

# \donttest{
cdm <- mockIncidencePrevalenceRef(
  sampleSize = 10000,
  earliestObservationStartDate = as.Date("2010-01-01"),
  latestObservationStartDate = as.Date("2018-01-01")
)
timings <- IncidencePrevalence::benchmarkIncidencePrevalence(cdm)
#> Loading required namespace: testthat
#> Creating denominator cohorts
#>  -- getting cohort dates for ■■■■■■■■■■■■■■■■                 4 of 8 cohorts
#>  -- getting cohort dates for ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  8 of 8 cohorts
#> Time taken to get cohorts: 0 min and 8 sec
#> Getting prevalence for analysis 1 of 8
#> Getting prevalence for analysis 2 of 8
#> Getting prevalence for analysis 3 of 8
#> Getting prevalence for analysis 4 of 8
#> Getting prevalence for analysis 5 of 8
#> Getting prevalence for analysis 6 of 8
#> Getting prevalence for analysis 7 of 8
#> Getting prevalence for analysis 8 of 8
#> Time taken: 0 mins and 8 secs
#> Getting prevalence for analysis 1 of 8
#> Getting prevalence for analysis 2 of 8
#> Getting prevalence for analysis 3 of 8
#> Getting prevalence for analysis 4 of 8
#> Getting prevalence for analysis 5 of 8
#> Getting prevalence for analysis 6 of 8
#> Getting prevalence for analysis 7 of 8
#> Getting prevalence for analysis 8 of 8
#> Time taken: 0 mins and 17 secs
#> Getting prevalence for analysis 1 of 8
#> Getting prevalence for analysis 2 of 8
#> Getting prevalence for analysis 3 of 8
#> Getting prevalence for analysis 4 of 8
#> Getting prevalence for analysis 5 of 8
#> Getting prevalence for analysis 6 of 8
#> Getting prevalence for analysis 7 of 8
#> Getting prevalence for analysis 8 of 8
#> Time taken: 0 mins and 11 secs
#> Getting prevalence for analysis 1 of 8
#> Getting prevalence for analysis 2 of 8
#> Getting prevalence for analysis 3 of 8
#> Getting prevalence for analysis 4 of 8
#> Getting prevalence for analysis 5 of 8
#> Getting prevalence for analysis 6 of 8
#> Getting prevalence for analysis 7 of 8
#> Getting prevalence for analysis 8 of 8
#> Time taken: 0 mins and 48 secs
#> Getting incidence for analysis 1 of 8
#> Getting incidence for analysis 2 of 8
#> Getting incidence for analysis 3 of 8
#> Getting incidence for analysis 4 of 8
#> Getting incidence for analysis 5 of 8
#> Getting incidence for analysis 6 of 8
#> Getting incidence for analysis 7 of 8
#> Getting incidence for analysis 8 of 8
#> Overall time taken: 0 mins and 7 secs
#> Getting incidence for analysis 1 of 8
#> Getting incidence for analysis 2 of 8
#> Getting incidence for analysis 3 of 8
#> Getting incidence for analysis 4 of 8
#> Getting incidence for analysis 5 of 8
#> Getting incidence for analysis 6 of 8
#> Getting incidence for analysis 7 of 8
#> Getting incidence for analysis 8 of 8
#> Overall time taken: 0 mins and 21 secs
# }