Collect population incidence estimates

Collect population incidence estimates

Usage

estimateIncidence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorCohortId = NULL,
  outcomeCohortId = NULL,
  interval = "years",
  completeDatabaseIntervals = TRUE,
  outcomeWashout = Inf,
  repeatedEvents = FALSE,
  minCellCount = 5,
  temporary = TRUE,
  returnParticipants = FALSE
)

Arguments

cdm

A CDM reference object

denominatorTable

A cohort table with a set of denominator cohorts (for example, created using the generateDenominatorCohortSet() function).

outcomeTable

A cohort table in the cdm reference containing a set of outcome cohorts.

denominatorCohortId

The cohort definition ids of the denominator cohorts of interest. If NULL all cohorts will be considered in the analysis.

outcomeCohortId

The cohort definition ids of the outcome cohorts of interest. If NULL all cohorts will be considered in the analysis.

interval

Time intervals over which incidence is estimated. Can be "weeks", "months", "quarters", "years", or "overall". ISO weeks will be used for weeks. Calendar months, quarters, or years can be used, or an overall estimate for the entire time period observed (from earliest cohort start to last cohort end) can also be estimated. If more than one option is chosen then results will be estimated for each chosen interval.

completeDatabaseIntervals

TRUE/ FALSE. Where TRUE, incidence will only be estimated for those intervals where the database captures all the interval (based on the earliest and latest observation period start dates, respectively).

outcomeWashout

The number of days used for a 'washout' period between the end of one outcome and an individual starting to contribute time at risk. If Inf, no time can be contributed after an event has occurred (whether during the study period or if occurring beforehand).

repeatedEvents

TRUE/ FALSE. If TRUE, an individual will be able to contribute multiple events during the study period (time while they are present in an outcome cohort and any subsequent washout will be excluded). If FALSE, an individual will only contribute time up to their first event during the study period.

minCellCount

The minimum number of events to reported, below which results will be obscured. If 0, all results will be reported.

temporary

If TRUE, temporary tables will be used throughout. If FALSE, permanent tables will be created in the write_schema of the cdm using the write_prefix (if specified). Note existing permanent tables in the write schema starting with the write_prefix will be at risk of being dropped or overwritten.

returnParticipants

Either TRUE or FALSE. If TRUE references to participants from the analysis will be returned allowing for further analysis. Note, if using permanent tables and returnParticipants is TRUE, one table per analysis will be kept in the cdm write schema.

Value

Incidence estimates

Examples

# \donttest{
cdm <- mockIncidencePrevalenceRef(sampleSize = 10000)
cdm <- generateDenominatorCohortSet(
  cdm = cdm, name = "denominator",
  cohortDateRange = c(as.Date("2008-01-01"), as.Date("2018-01-01"))
)
#> Creating denominator cohorts
#> Time taken to get cohorts: 0 min and 3 sec
inc <- estimateIncidence(
  cdm = cdm,
  denominatorTable = "denominator",
  outcomeTable = "outcome"
)
#> Getting incidence for analysis 1 of 1
#> Overall time taken: 0 mins and 1 secs
# }