Identify a set of denominator populations

generateDenominatorCohortSet() creates a set of cohorts that can be used for the denominator population in analyses of incidence, using estimateIncidence(), or prevalence, using estimatePointPrevalence() or estimatePeriodPrevalence().

Usage

generateDenominatorCohortSet(
  cdm,
  name,
  cohortDateRange = NULL,
  ageGroup = list(c(0, 150)),
  sex = "Both",
  daysPriorObservation = 0,
  requirementInteractions = TRUE,
  targetCohortTable = NULL,
  targetCohortId = NULL,
  overwrite = FALSE
)

Arguments

cdm

A CDM reference object

name

Name of the cohort table to be created.

cohortDateRange

Two dates. The first indicating the earliest cohort start date and the second indicating the latest possible cohort end date. If NULL or the first date is set as missing, the earliest observation_start_date in the observation_period table will be used for the former. If NULL or the second date is set as missing, the latest observation_end_date in the observation_period table will be used for the latter.

ageGroup

A list of age groups for which cohorts will be generated. A value of list(c(0,17), c(18,30)) would, for example, lead to the creation of cohorts for those aged from 0 to 17, and from 18 to 30. In this example an individual turning 18 during the time period would appear in both cohorts (leaving the first cohort the day before their 18th birthday and entering the second from the day of their 18th birthday).

sex

Sex of the cohorts. This can be one or more of: "Male", "Female", or "Both".

daysPriorObservation

The number of days of prior observation observed in the database required for an individual to start contributing time in a cohort.

requirementInteractions

If TRUE, cohorts will be created for all combinations of ageGroup, sex, and daysPriorObservation. If FALSE, only the first value specified for the other factors will be used. Consequently, order of values matters when requirementInteractions is FALSE.

targetCohortTable

A cohort table in the cdm reference to use to limit cohort entry and exit (with individuals only contributing to a cohort when they are contributing to the cohort in the target table).

targetCohortId

The cohort definition id for the cohort of interest in the target table. If targetCohortTable is specified, a single targetCohortId must also be specified.

overwrite

Whether to overwrite any existing table with the same name

Value

A cohort reference

Examples

# \donttest{
cdm <- mockIncidencePrevalenceRef(sampleSize = 10000)
cdm <- generateDenominatorCohortSet(
  cdm = cdm,
  name = "denominator",
  cohortDateRange = as.Date(c("2008-01-01", "2020-01-01"))
)
#> ℹ Creating denominator cohorts
#> ✔ Cohorts created in 0 min and 3 sec
cdm$denominator
#> # Source:   table<main.denominator> [?? x 4]
#> # Database: DuckDB v0.9.1 [unknown@Linux 6.2.0-1014-azure:R 4.3.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int> <chr>      <date>            <date>         
#>  1                    1 3          2008-01-01        2008-04-02     
#>  2                    1 4          2009-06-19        2010-01-03     
#>  3                    1 6          2012-10-31        2012-11-12     
#>  4                    1 7          2011-03-29        2012-09-30     
#>  5                    1 8          2016-09-11        2018-04-03     
#>  6                    1 9          2019-02-21        2020-01-01     
#>  7                    1 10         2015-02-05        2017-01-13     
#>  8                    1 11         2018-12-04        2020-01-01     
#>  9                    1 12         2016-02-19        2017-04-09     
#> 10                    1 13         2008-01-01        2009-03-30     
#> # ℹ more rows
# }